Add test tracing and test trace storage, refactor Playwright test class name and test name
